migrate URI sidepanel
