+++ frontend/packages/knative-plugin/src/components/overview/__tests__/KnativeResourceOverviewPage.spec.tsx
@@ -9,9 +9,7 @@ import {
   revisionObj,
   EventSubscriptionObj,
 } from '../../../topology/__tests__/topology-knative-test-data';
-import { URI_KIND } from '../../../topology/const';
 import { KnativeResourceOverviewPage } from '../KnativeResourceOverviewPage';
-import SinkUriResourcesTab from '../SinkUriResourcesTab';
 
 describe('KnativeResourceOverviewPage', () => {
   let item: OverviewItem;
@@ -54,22 +52,5 @@ describe('KnativeResourceOverviewPage', () => {
     expect(resourceOverviewDetails).toHaveLength(1);
     expect(resourceOverviewDetails.at(0).props().menuActions).toHaveLength(4);
     expect(resourceOverviewDetails.at(0).props().kindObj).toEqual(EventingSubscriptionModel);
-  });
-
-  it('should render SinkUriResourcesTab if NodeType is sink-uri', () => {
-    const itemData = {
-      ...item,
-      obj: {
-        kind: URI_KIND,
-        metadata: {
-          uid: '02c34a0e-9638-11e9-b134-06a61d886b62_nodesinkuri',
-        },
-        spec: { sinkUri: 'http://overlayimage.testproject3.svc.cluster.local' },
-      },
-    };
-    const wrapper = shallow(
-      <KnativeResourceOverviewPage item={itemData} knativeModels={[RevisionModel]} kindsInFlight />,
-    );
-    expect(wrapper.find(SinkUriResourcesTab)).toHaveLength(1);
   });
 });
